+++ packages/eslint-plugin/src/rules/no-restricted-imports.ts
@@ -146,7 +146,13 @@ export default createRule<Options, MessageIds>({
         typeof restrictedPattern === 'object' &&
         restrictedPattern.allowTypeImports
       ) {
-        allowedImportTypeMatchers.push(ignore().add(restrictedPattern.group));
+        // Following how ignore is configured in the base rule
+        allowedImportTypeMatchers.push(
+          ignore({
+            allowRelativePaths: true,
+            ignoreCase: !restrictedPattern.caseSensitive,
+          }).add(restrictedPattern.group),
+        );
       }
     }
     function isAllowedTypeImportPattern(importSource: string): boolean {