+++ packages/eslint-plugin/typings/eslint-rules.d.ts
@@ -416,6 +416,7 @@ declare module 'eslint/lib/rules/no-unused-vars' {
           argsIgnorePattern?: string;
           caughtErrors?: 'all' | 'none';
           caughtErrorsIgnorePattern?: string;
+          destructuredArrayIgnorePattern?: string;
         },
     ],
     {
@@ -842,6 +843,29 @@ declare module 'eslint/lib/rules/space-infix-ops' {
   export = rule;
 }
 
+declare module 'eslint/lib/rules/space-before-blocks' {
+  import { TSESLint, TSESTree } from '@typescript-eslint/utils';
+
+  const rule: TSESLint.RuleModule<
+    'missingSpace' | 'unexpectedSpace',
+    [
+      | 'always'
+      | 'never'
+      | {
+          classes?: 'always' | 'never' | 'off';
+          functions?: 'always' | 'never' | 'off';
+          keywords?: 'always' | 'never' | 'off';
+        },
+    ],
+    {
+      BlockStatement(node: TSESTree.BlockStatement): void;
+      ClassBody(node: TSESTree.ClassBody): void;
+      SwitchStatement(node: TSESTree.SwitchStatement): void;
+    }
+  >;
+  export = rule;
+}
+
 declare module 'eslint/lib/rules/prefer-const' {
   import { TSESLint, TSESTree } from '@typescript-eslint/utils';
 
@@ -905,6 +929,7 @@ declare module 'eslint/lib/rules/no-restricted-imports' {
       | {
           group: string[];
           message?: string;
+          caseSensitive?: boolean;
           // extended
           allowTypeImports?: boolean;
         }[];