+++ packages/scope-manager/src/lib/es2022.error.ts
@@ -1,0 +1,19 @@
+// THIS CODE WAS AUTOMATICALLY GENERATED
+// DO NOT EDIT THIS CODE BY HAND
+// YOU CAN REGENERATE IT USING yarn generate:lib
+
+import { ImplicitLibVariableOptions } from '../variable';
+import { TYPE } from './base-config';
+
+export const es2022_error = {
+  ErrorOptions: TYPE,
+  Error: TYPE,
+  ErrorConstructor: TYPE,
+  EvalErrorConstructor: TYPE,
+  RangeErrorConstructor: TYPE,
+  ReferenceErrorConstructor: TYPE,
+  SyntaxErrorConstructor: TYPE,
+  TypeErrorConstructor: TYPE,
+  URIErrorConstructor: TYPE,
+  AggregateErrorConstructor: TYPE,
+} as Record<string, ImplicitLibVariableOptions>;