+++ packages/type-utils/tests/isTypeReadonly.test.ts
@@ -82,6 +82,14 @@ describe('isTypeReadonly', () => {
             ['type Test = Readonly<ReadonlySet<string>>;'],
             ['type Test = Readonly<ReadonlyMap<string, string>>;'],
           ])('handles fully readonly sets and maps', runTests);
+
+          // Private Identifier.
+          // Note: It can't be accessed from outside of class thus exempt from the checks.
+          it.each([
+            ['class Foo { readonly #readonlyPrivateField = "foo"; }'],
+            ['class Foo { #privateField = "foo"; }'],
+            ['class Foo { #privateMember() {}; }'],
+          ])('treat private identifier as readonly', runTests);
         });
 
         describe('is not readonly', () => {