+++ packages/utils/src/ast-utils/index.ts
@@ -1,3 +1,4 @@
+export * from './eslint-utils';
+export * from './helpers';
 export * from './misc';
 export * from './predicates';
-export * from './eslint-utils';